requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.1
langchain==0.1.4
langchain-openai==0.0.5
openai==1.12.0
supabase==2.3.1
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
streamlit==1.32.0 